OnB-2-1 fixed Float

+++ src/main/kotlin/lesson_2/L2_task1.kt
@@ -5,10 +5,10 @@ fun main() {
     val studentMark2: Byte = 4
     val studentMark3: Byte = 3
     val studentMark4: Byte = 5
-    val numberOfStudents: Float = 4f
+    val numberOfStudents: Byte = 4
 
     //Вычисление среднего балла
-    val avgRating: Float = (studentMark1 + studentMark2 + studentMark3 + studentMark4) / numberOfStudents
+    val avgRating: Float = (studentMark1 + studentMark2 + studentMark3 + studentMark4).toFloat() / numberOfStudents
 
     println(String.format("%.2f", avgRating))
 }
